% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FishMaps.R
\docType{data}
\name{BB.data.y}
\alias{BB.data.y}
\title{Baitboat yearly aggregated data}
\format{A data frame with 56 observations on the following 4 variables.
\describe{ \item{list("year")}{a factor with levels \code{2001}
\code{2002}} \item{list("lat")}{a numeric vector} \item{list("lon")}{a
numeric vector} \item{list("cpue")}{a numeric vector} }}
\source{
Randomly generated data.
}
\usage{
data(BB.data.y)
}
\description{
Skipjack tuna CPUE by year, caught by the brazilian baitboat fleet, based
at Itajai (SC) harbor.
}
\examples{
data(BB.data.y)
str(BB.data.y)
}
\keyword{datasets}

