% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FishMaps.R
\docType{data}
\name{BB.data.yq}
\alias{BB.data.yq}
\title{Baitboat quarterly aggregated data}
\format{A data frame with 120 observations on the following 5 variables:
\itemize{
  \item year: a factor with levels \code{2001}, \code{2002}
  \item quarter: a factor with levels \code{1}, \code{2}, \code{3}, \code{4}
  \item lat: a numeric vector
  \item lon: a numeric vector
  \item cpue: a numeric vector
}}
\source{
Actually this is some randomly generated data.
}
\usage{
data(BB.data.yq)
}
\description{
Skipjack tuna CPUE by quarter and year, caught by the brazilian baitboat
fleet, based at Itajai (SC) harbor.
}
\examples{
data(BB.data.yq)
str(BB.data.yq)
}
\keyword{datasets}

