% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FishMaps.R
\docType{package}
\name{FishMaps}
\alias{FishMaps}
\alias{FishMaps-package}
\title{Proportional symbol mapping for fishery data in batch mode}
\description{
Proportional symbol mapping for fishery data in batch
mode (lattice version)
}

