% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishMaps.R
\docType{data}
\name{LL.data.y}
\alias{LL.data.y}
\title{Longline yearly aggregated data}
\format{A data frame with 82 observations on the following 4 variables.
\itemize{
  \item year: a factor with levels \code{2001}, \code{2002},
    \code{2003}, \code{2004}, \code{2005}
  \item lat: a numeric vector
  \item lon: a numeric vector
  \item cpue: a numeric vector
}}
\source{
Actually this is some randomly generated data.
}
\usage{
data(LL.data.y)
}
\description{
Swordfish CPUE by year, caught by the brazilian longline fleet, based at
Itajai (SC) harbor.
}
\examples{

data(LL.data.y)
str(LL.data.y)

}
\keyword{datasets}

