% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FishMaps.R
\docType{data}
\name{LL.data.yq}
\alias{LL.data.yq}
\title{Longline quarterly aggregated data}
\format{A data frame with 181 observations on the following 5 variables.
\describe{ \item{list("year")}{a factor with levels \code{2001} \code{2002}
\code{2003} \code{2004} \code{2005}} \item{list("quarter")}{a factor with
levels \code{1} \code{2} \code{3} \code{4}} \item{list("lat")}{a numeric
vector} \item{list("lon")}{a numeric vector} \item{list("cpue")}{a numeric
vector} }}
\source{
Randomly generated data.
}
\usage{
data(LL.data.yq)
}
\description{
Swordfish CPUE by year and quarter, caught by the brazilian longline fleet,
based at Itajai (SC) harbor.
}
\examples{
data(LL.data.yq)
    str(LL.data.yq)
}
\keyword{datasets}

