% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.fishmaps.R
\name{panel.fishmaps}
\alias{panel.fishmaps}
\title{A lattice panel function}
\usage{
panel.fishmaps(x, y, z, map.db, msq, breaks, col.reg, labsx, labsy, subscripts,
  ...)
}
\arguments{
\item{x}{from lattice fomrmula x}

\item{y}{from lattice fomrmula y}

\item{z}{from lattice fomrmula z}

\item{map.db}{map database}

\item{msq}{middle square}

\item{breaks}{from main function}

\item{col.reg}{from main function}

\item{labsx}{from main function}

\item{labsy}{from main function}

\item{subscripts}{lattice subscripts}

\item{...}{other arguments passed to other lattice panels}
}
\value{
Nothing
}
\description{
A lattice panel function to draw rectangles where z is
different from zero and to draw points where z is exactly zero
}
\author{
Fernando Mayer
}

