% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.zero.points.R
\name{panel.zero.points}
\alias{panel.zero.points}
\title{A lattice panel function}
\usage{
panel.zero.points(x, y, z, subscripts, ...)
}
\arguments{
\item{x}{from lattice fomrmula x}

\item{y}{from lattice fomrmula y}

\item{z}{from lattice fomrmula z}

\item{subscripts}{lattice subscripts}

\item{...}{other arguments passed to other lattice panels}
}
\value{
Nothing.
}
\description{
A lattice panel function to plot special characters to
maps when data is zero
}
\author{
Fernando Mayer
}

