% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointmap.R
\name{pointmap}
\alias{pointmap}
\title{Plots Points into Maps.}
\source{
The databases \code{"world"} and \code{"worldHires"} were
extracted from the databases of the same names in packages
\code{maps} and \code{mapdata}, respectively. These databases are
from the CIA World Data Bank II, which are in public domain and
currently (mid-2003) available at
\url{http://www.evl.uic.edu/pape/data/WDB}.
}
\usage{
pointmap(
  x,
  data,
  xlim,
  ylim,
  square = 1,
  col.land = "snow",
  pch.points = 20,
  col.points = "black",
  cex.points = 1,
  database = c("world", "worldHires"),
  ...
)
}
\arguments{
\item{x}{A lattice fomula interface, usually \code{y ~ x} where
x are the longitude points, and y are the latitude points.}

\item{data}{A data frame where the variables in the formula
interface (\code{x}) are contained.}

\item{xlim}{The x limits for the map. Usually this will be the
longitude limits.}

\item{ylim}{The y limits for the map. Usually this will be the
latitude limits.}

\item{square}{Size of the square. It must be a length one numeric
vector (e.g. for squares of 1x1 degrees, this should be 1). Roughly
speaking this can be viewed as the resolution of the map.}

\item{col.land}{The color of land in map. Any valid color in R is
accepted, see \code{\link[grDevices]{colors}}. By default it's
\code{"snow"}, a pale light color.}

\item{pch.points}{The _p_oint _ch_aracter to be plotted.}

\item{col.points}{The color of the points to be plotted.}

\item{cex.points}{The size of the points to be plotted.}

\item{database}{The maps database to be used to plot the
coastlines. Can be one of \code{"world"} (lower resolution) or
\code{"worldHires"} (higher resolution). Defaults to
\code{"world"}.}

\item{...}{Other arguments passed to
\code{\link[lattice]{levelplot}}.}
}
\value{
A figure with the map(s) and the data points.
}
\description{
Plots georeferenced data (e.g. sampling points) into
    maps.
}
\details{
Made to plot sampling points in the ocean.
}
\seealso{
\code{\link[lattice]{levelplot}}, \code{\link[lattice]{xyplot}}
}
\author{
Fernando Mayer \email{fernandomayer@gmail.com}
}
