% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xyticks.R
\name{xyticks}
\alias{xyticks}
\title{Creates the ticks for the maps}
\usage{
xyticks(xlim, ylim, square)
}
\arguments{
\item{xlim}{x limits}

\item{ylim}{y limits}

\item{square}{size of square}
}
\value{
A list with 4 components, to be used in levelmap.
}
\description{
Creates the ticks and labels to print in the maps. This
is an internal function.
}
\author{
Fernando Mayer
}

