% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishMaps.R
\docType{data}
\name{BB.data.y}
\alias{BB.data.y}
\title{Baitboat yearly aggregated data}
\format{A data frame with 56 observations on the following 4 variables:
\itemize{
  \item year: a factor with levels \code{2001}, \code{2002}
  \item lat: a numeric vector
  \item lon: a numeric vector
  \item cpue: a numeric vector
}}
\source{
Actually this is some randomly generated data.
}
\usage{
data(BB.data.y)
}
\description{
Skipjack tuna CPUE by year, caught by the brazilian baitboat fleet, based
at Itajai (SC) harbor.
}
\examples{

data(BB.data.y)
str(BB.data.y)

}
\keyword{datasets}

