% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FishMaps.R
\docType{package}
\name{FishMaps}
\alias{FishMaps}
\alias{FishMaps-package}
\title{Plots Fishery Data into Maps.}
\description{
Plots georeferenced fishery data (e.g. catch, effort
and CPUE) into maps. This is the lattice version of a previous
FishMaps version based on traditional grid graphics.
}

