% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/levelmap.R
\name{levelmap}
\alias{levelmap}
\title{Plots fishery data into maps}
\usage{
levelmap(x, data, xlim, ylim, breaks, jump, key.space = "right",
  database = "world", bathymetry = FALSE, bathymetry.seq = NULL, ...)
}
\arguments{
\item{x}{}

\item{data}{}

\item{xlim}{}

\item{ylim}{}

\item{breaks}{}

\item{jump}{}

\item{key.space}{}

\item{database}{}

\item{bathymetry}{}

\item{bathymetry.seq}{}

\item{...}{}
}
\value{
A map with data
}
\description{
Plots georeferenced fishery data into maps. This
desciption must be expanded.
}
\author{
Fernando Mayer
}

