% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/levelmap.R
\name{levelmap}
\alias{levelmap}
\title{Plots fishery data into maps}
\usage{
levelmap(x, data, xlim, ylim, breaks, square = 1, key.space = "right",
  database = c("world", "worldHires"))
}
\arguments{
\item{x}{fomula interface}

\item{data}{data frame}

\item{xlim}{x limits}

\item{ylim}{y limits}

\item{breaks}{class breaks}

\item{square}{size of square}

\item{key.space}{place for the legend}

\item{database}{world or worldHires}
}
\value{
A map with data
}
\description{
Plots georeferenced fishery data into maps. This
desciption must be expanded.
}
\author{
Fernando Mayer
}

