% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.fishmaps.R
\name{panel.fishmaps}
\alias{panel.fishmaps}
\title{A lattice panel function}
\usage{
panel.fishmaps(x, y, z, map.db, msq, breaks, col.reg, labsx, labsy, subscripts,
  ...)
}
\arguments{
\item{x}{From lattice formula x.}

\item{y}{From lattice formula y.}

\item{z}{From lattice formula z.}

\item{map.db}{The map database.}

\item{msq}{The middle of squares defined in argument \code{square}
in \code{\link{levelmap}}.}

\item{breaks}{The \code{breaks} argument from \code{\link{levelmap}}.}

\item{col.reg}{The color regions, from \code{\link{levelmap}}.}

\item{labsx}{The x axis ticks and labels.}

\item{labsy}{The y axis ticks and labels.}

\item{subscripts}{Lattice panel subscripts.}

\item{...}{Other arguments passed to lattice panel functions.}
}
\description{
A lattice panel function to draw rectangles where z is
different from zero and to draw points where z is exactly zero.
}
\details{
All arguments used here are defined in the source of the
\code{\link{levelmap}} function.
}
\author{
Fernando Mayer \email{fernandomayer@gmail.com}
}

