% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.zero.points.R
\name{panel.zero.points}
\alias{panel.zero.points}
\title{A lattice panel function}
\usage{
panel.zero.points(x, y, z, subscripts, ...)
}
\arguments{
\item{x}{From lattice fomrmula x.}

\item{y}{From lattice fomrmula y.}

\item{z}{From lattice fomrmula z.}

\item{subscripts}{Lattice panbel subscripts.}

\item{...}{Other arguments passed to lattice panel functions.}
}
\description{
A lattice panel function to plot special characters to
maps when data is zero.
}
\author{
Fernando Mayer \email{fernandomayer@gmail.com}
}

