% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyticks.R
\name{xyticks}
\alias{xyticks}
\title{Creates the ticks for the maps.}
\usage{
xyticks(xlim, ylim, square)
}
\arguments{
\item{xlim}{The x limits for the map. Usually this will be the
longitude limits.}

\item{ylim}{The y limits for the map. Usually this will be the
latitude limits.}

\item{square}{Size of the square. It must be a length one numeric
vector (e.g. for squares of 1x1 degrees, this should be 1). Roughly
speaking this can be viewed as the resolution of the map.}
}
\value{
A list with 4 components, to be used in levelmap.
}
\description{
Creates the ticks and labels to print in the maps. This
is an internal function.
}
\author{
Fernando Mayer \email{fernandomayer@gmail.com}
}
